New Effects
Added a new effects system.
Works off the same principal as the entity system, but I needed
different features for the Effects system.
This was used to add the stars in the background
The render call is split into two parts, front and back. This allows
things to be rendered either in front or behind objects depending on
it's Z location set in the initializer method

+++ Neon/src/net/toper/graphics/Render.java
@@ -17,7 +17,7 @@ public class Render {
 		img.draw((int) x, (int) y, scale);
 	}
 
-	public void drawRect(float x, float y, float width, int height, Color c) {
+	public void drawRect(float x, float y, float width, float height, Color c) {
 		Main.g.setColor(c);
 		Main.g.fillRect(x, y, width, height);
 	}
